requirements.txt
fastapi
uvicorn
tensorflow
keras
opencv-python
faiss-cpu
pillow
numpy
scikit-learn
tqdm
jinja2>=3.0
python-multipart